LayoutTests/imported/w3c/web-platform-tests/event-timing/event-retarget-expected.txt


FAIL Event Timing: target reports the last Event Target, i.e. nothing from shadow DOM. assert_equals: There should be a single first-input entry expected 1 but got 0
